обновлёно активити поиска
