+++ app/src/main/res/layout/search_service.xml
@@ -39,6 +39,7 @@
         android:id="@+id/citySearchServiceSpinner"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
+        android:background="@color/yellow1"
         app:layout_constraintBottom_toTopOf="@+id/footerSearchServiceLayout"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.1"
@@ -50,6 +51,7 @@
         android:id="@+id/searchBySearchServiceSpinner"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
+        android:background="@color/yellow1"
         app:layout_constraintBottom_toTopOf="@+id/footerSearchServiceLayout"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.6"
